feat: add person and tv show features

+++ core/data/src/main/java/core/data/account/AccountRepositoryImpl.kt
@@ -9,17 +9,24 @@ import com.github.michaelbull.result.onSuccess
 import core.data.SessionManager
 import core.data.account.model.AccountDetailsResponse
 import core.data.account.model.FavoriteMovieRequest
+import core.data.account.model.FavoriteTvShowDto
 import core.data.account.model.WatchlistMovieRequest
+import core.data.account.model.WatchlistTvShowRequest
 import core.data.account.model.toDto
 import core.data.common.model.MovieDto
+import core.data.common.model.TvShowDto
 import core.data.network.toFailure
 import core.database.LanguagePreference
 import core.database.account.FavoriteMovieDao
 import core.database.account.FavoriteTvShowDao
 import core.database.account.RatedMovieDao
+import core.database.account.RatedTvShowDao
 import core.database.account.WatchlistMovieDao
+import core.database.account.WatchlistTvShowDao
 import core.database.account.model.FavoriteMovieIdEntity
+import core.database.account.model.FavoriteTvShowIdEntity
 import core.database.account.model.WatchlistMovieIdEntity
+import core.database.account.model.WatchlistTvShowIdEntity
 import core.database.user.UserDetails
 import core.database.user.UserManager
 import core.domain.AuthFailure
@@ -40,6 +47,8 @@ class AccountRepositoryImpl(
     private val watchlistMovieDao: WatchlistMovieDao,
     private val ratedMovieDao: RatedMovieDao,
     private val favoriteTvShowDao: FavoriteTvShowDao,
+    private val watchlistTvShowDao: WatchlistTvShowDao,
+    private val ratedTvShowDao: RatedTvShowDao,
     private val languagePreference: LanguagePreference,
     private val strResource: StrResource,
 ) : AccountRepository {
@@ -98,6 +107,39 @@ class AccountRepositoryImpl(
         }
     }
 
+    override suspend fun fetchFavoriteTvShows(sortBy: SortBy.CreatedAt) {
+        withContext(Dispatchers.IO) {
+            favoriteTvShowDao.clearAllIds()
+            val favoriteTvShows = mutableListOf<TvShowDto>()
+
+            var endPage = 2
+            var page = 1
+            while (page in 0..endPage) {
+                accountService.getFavoriteTvShows(
+                    page,
+                    sessionManager.requireSessionId(),
+                    languagePreference.getLanguageTag(),
+                    sortBy.toDto().value
+                ).onSuccess { response ->
+                    endPage = response.totalPages
+
+                    if (response.results.isNotEmpty()) {
+                        favoriteTvShows.addAll(response.results)
+                    }
+                }.onFailure {
+                    break
+                }
+                page++
+            }
+
+            favoriteTvShowDao.insertIds(
+                *favoriteTvShows
+                    .map { it.toFavoriteTvShowIdEntity() }
+                    .toTypedArray()
+            )
+        }
+    }
+
     override suspend fun fetchWatchlistMovies(sortBy: SortBy.CreatedAt) {
         withContext(Dispatchers.IO) {
             watchlistMovieDao.clearAllIds()
@@ -132,6 +174,40 @@ class AccountRepositoryImpl(
         }
     }
 
+    override suspend fun fetchWatchlistTvShows(sortBy: SortBy.CreatedAt) {
+        withContext(Dispatchers.IO) {
+            watchlistTvShowDao.clearAllIds()
+            val watchlistTvShows = mutableListOf<TvShowDto>()
+
+            var endPage = 2
+            var page = 1
+            while (page in 0..endPage) {
+                accountService.getWatchlistTvShows(
+                    page,
+                    sessionManager.requireSessionId(),
+                    languagePreference.getLanguageTag(),
+                    sortBy.toDto().value
+
+                ).onSuccess { response ->
+                    endPage = response.totalPages
+
+                    if (response.results.isNotEmpty()) {
+                        watchlistTvShows.addAll(response.results)
+                    }
+                }.onFailure {
+                    break
+                }
+                page++
+            }
+
+            watchlistTvShowDao.insertIds(
+                *watchlistTvShows
+                    .map { it.toWatchlistTvShowIdEntity() }
+                    .toTypedArray()
+            )
+        }
+    }
+
     override suspend fun syncMovieFavoriteStatus(
         movieId: Int,
         isFavorite: Boolean,
@@ -157,7 +233,12 @@ class AccountRepositoryImpl(
                     Ok(Unit)
                 },
                 failure = { networkError ->
-                    favoriteMovieDao.deleteIds(FavoriteMovieIdEntity(movieId))
+                    if (isFavorite) {
+                        favoriteMovieDao.deleteIds(FavoriteMovieIdEntity(movieId))
+                    } else {
+                        favoriteMovieDao.insertIds(FavoriteMovieIdEntity(movieId))
+                    }
+
                     Err(networkError.toFailure())
                 }
             )
@@ -186,7 +267,75 @@ class AccountRepositoryImpl(
                     Ok(Unit)
                 },
                 failure = { networkError ->
-                    watchlistMovieDao.deleteIds(WatchlistMovieIdEntity(movieId))
+                    if (inWatchlist) {
+                        watchlistMovieDao.deleteIds(WatchlistMovieIdEntity(movieId))
+                    } else {
+                        watchlistMovieDao.insertIds(WatchlistMovieIdEntity(movieId))
+                    }
+                    Err(networkError.toFailure())
+                }
+            )
+        }
+    }
+
+    override suspend fun syncTvShowFavoriteStatus(
+        tvShowId: Int,
+        isFavorite: Boolean,
+    ): Result<Unit, Failure> {
+        return withContext(Dispatchers.IO) {
+            if (isFavorite) {
+                favoriteTvShowDao.insertIds(FavoriteTvShowIdEntity(tvShowId))
+            } else {
+                favoriteTvShowDao.deleteIds(FavoriteTvShowIdEntity(tvShowId))
+            }
+            val sessionId = sessionManager.getSessionId()
+                ?: return@withContext Err(AuthFailure(strResource.get(R.string.something_is_wrong)))
+
+            return@withContext accountService.postFavoriteTvShow(
+                FavoriteTvShowDto(
+                    favorite = isFavorite,
+                    mediaId = tvShowId
+                ),
+                sessionId = sessionId
+            ).mapBoth(
+                success = { Ok(Unit) },
+                failure = { networkError ->
+                    if (isFavorite) {
+                        favoriteTvShowDao.deleteIds(FavoriteTvShowIdEntity(tvShowId))
+                    } else {
+                        favoriteTvShowDao.insertIds(FavoriteTvShowIdEntity(tvShowId))
+                    }
+                    Err(networkError.toFailure())
+                }
+            )
+        }
+    }
+
+    override suspend fun syncTvShowWatchlistStatus(
+        tvShowId: Int,
+        inWatchlist: Boolean,
+    ): Result<Unit, Failure> {
+        return withContext(Dispatchers.IO) {
+            if (inWatchlist) {
+                watchlistTvShowDao.insertIds(WatchlistTvShowIdEntity(tvShowId))
+            } else {
+                watchlistTvShowDao.deleteIds(WatchlistTvShowIdEntity(tvShowId))
+            }
+
+            return@withContext accountService.postWatchlistTvShow(
+                WatchlistTvShowRequest(
+                    watchlist = inWatchlist,
+                    mediaId = tvShowId
+                ),
+                sessionId = sessionManager.requireSessionId()
+            ).mapBoth(
+                success = { Ok(Unit) },
+                failure = { networkError ->
+                    if (inWatchlist) {
+                        watchlistTvShowDao.deleteIds(WatchlistTvShowIdEntity(tvShowId))
+                    } else {
+                        watchlistTvShowDao.insertIds(WatchlistTvShowIdEntity(tvShowId))
+                    }
                     Err(networkError.toFailure())
                 }
             )
@@ -199,8 +348,12 @@ class AccountRepositoryImpl(
             favoriteMovieDao.clearAllMovies()
             watchlistMovieDao.clearAllIds()
             watchlistMovieDao.clearAllMovies()
-            ratedMovieDao.clearAllMovies()
-            favoriteTvShowDao.clearAll()
+            ratedMovieDao.clearAll()
+            favoriteTvShowDao.clearAllTvShows()
+            favoriteTvShowDao.clearAllIds()
+            watchlistTvShowDao.clearAllTvShows()
+            watchlistTvShowDao.clearAllIds()
+            ratedTvShowDao.clearAll()
             userManager.removeCurrentUser()
         }
     }
@@ -212,6 +365,14 @@ private fun MovieDto.toFavoriteMovieIdEntity(): FavoriteMovieIdEntity {
 
 private fun MovieDto.toWatchlistMovieIdEntity(): WatchlistMovieIdEntity {
     return WatchlistMovieIdEntity(id)
+}
+
+private fun TvShowDto.toFavoriteTvShowIdEntity(): FavoriteTvShowIdEntity {
+    return FavoriteTvShowIdEntity(id)
+}
+
+private fun TvShowDto.toWatchlistTvShowIdEntity(): WatchlistTvShowIdEntity {
+    return WatchlistTvShowIdEntity(id)
 }
 
 private fun UserDetails.toUser(): User {
